+++ jdaviz/configs/mosviz/helper.py
@@ -464,6 +464,16 @@ class Mosviz(ConfigHelper, LineListMixin):
             self.load_metadata(spectra_2d, spectra=True)
             self.load_metadata(spectra_1d, spectra=True, sp1d=True, ids=spectra_1d)
 
+        elif spectra_1d and images:
+            self.load_1d_spectra(spectra_1d, spectra_1d_label)
+            self.load_images(images, images_label)
+            allow_link_table = False
+
+        elif spectra_2d and images:
+            self.load_2d_spectra(spectra_2d, spectra_2d_label)
+            self.load_images(images, images_label)
+            allow_link_table = False
+
         elif spectra_1d:
             self.load_1d_spectra(spectra_1d, spectra_1d_label)
             allow_link_table = False
@@ -473,14 +483,10 @@ class Mosviz(ConfigHelper, LineListMixin):
             allow_link_table = False
 
         else:
-            msg = "Warning: Please set valid values for the load_data() method"
-            msg = SnackbarMessage(msg, color='warning', sender=self)
-            self.app.hub.broadcast(msg)
+            self.app.hub.broadcast(SnackbarMessage(
+                "Warning: Please set valid values for the load_data() method",
+                color='warning', sender=self))
             return
-
-        if msg:
-            msg = SnackbarMessage(msg, color='warning', sender=self)
-            self.app.hub.broadcast(msg)
 
         if allow_link_table:
             self.link_table_data(None)